Added add champ button when login
