+++ src/components/SelectedChamp.jsx
@@ -4,12 +4,10 @@ import { useDispatch, useSelector } from 'react-redux';
 import Loading from './Loading';
 import Champ from './Champ';
 import { loadselected } from '../store/books';
-// import { selectIsLoggedIn } from '../store/sessionSlice';
 
 const SelectedChamp = ({ match }) => {
   const ID = match.params.id;
   const char = useSelector((state) => state.entities.champs.selected);
-  // const loggedIn = useSelector(selectIsLoggedIn);
 
   const dispatch = useDispatch();
