+++ src/store/books.js
@@ -8,7 +8,6 @@ export const champSlice = createSlice({
   initialState: {
     loading: false,
     list: {},
-    champSelection: '',
     champTags: '',
     category: 'All',
     selected: {},
@@ -23,9 +22,6 @@ export const champSlice = createSlice({
     },
     champsRequestFailed: (champs) => {
       champs.loading = false;
-    },
-    champSelection: (champs, action) => {
-      champs.champSelection = action.payload;
     },
     champCategory: (champs, action) => {
       champs.category = action.payload;
@@ -61,7 +57,8 @@ export const {
 export default champSlice.reducer;
 
 export const selectCategory = (state) => state.entities.champs.category;
-export const selectChamp = (state) => state.entities.champs.champSelection;
+export const selectChamp = (state) => state.entities.champs.selected;
+export const selectTag = (state) => state.entities.champs.champTags;
 
 const url = '/champs';
